Address review feedback (Greptile)
- .prettierrc: printWidth 110 → 120 (Raycast standard).
- preferences.ts: drop manual Zod schema, use auto-generated Preferences
  type from raycast-env.d.ts.
- workspace.ts: remove dead path-traversal guard (path.resolve already
  normalises .. segments; preference value is user-controlled).
- format.ts: extract formatRelative helper, deduplicate from
  resume-session.tsx and flagged-sessions.tsx.
- package.json: rename "New Session (Quick)" → "Quick New Session" to
  satisfy Title Case rule.
- package-lock.json: regenerated under Node 20 to match Raycast CI.

Note: metadata/ folder with screenshots is still pending — will follow
up once screenshots are captured.
