+++ extensions/craft-agents/.prettierrc
@@ -1,5 +1,5 @@
 {
-  "printWidth": 110,
+  "printWidth": 120,
   "singleQuote": false,
   "trailingComma": "all",
   "semi": true,
