+++ extensions/craft-agents/src/resume-session.tsx
@@ -1,23 +1,11 @@
 import { Action, ActionPanel, Icon, List, open, showToast, Toast } from "@raycast/api";
 import { showFailureToast, useCachedPromise } from "@raycast/utils";
 import { buildAction, buildSessionLink } from "./lib/deeplink";
+import { formatRelative } from "./lib/format";
 import { getPrefs } from "./lib/preferences";
 import { listSessions } from "./lib/sessions";
 import type { SessionRecord } from "./lib/sessions.schema";
 import { logger } from "./lib/logger";
-
-function formatRelative(date: Date | undefined): string {
-  if (!date) return "";
-  const diff = Date.now() - date.getTime();
-  const m = Math.floor(diff / 60_000);
-  if (m < 1) return "just now";
-  if (m < 60) return `${m}m ago`;
-  const h = Math.floor(m / 60);
-  if (h < 24) return `${h}h ago`;
-  const d = Math.floor(h / 24);
-  if (d < 30) return `${d}d ago`;
-  return date.toLocaleDateString();
-}
 
 function subtitle(rec: SessionRecord): string {
   const parts: string[] = [];
@@ -77,11 +65,7 @@ export default function Command() {
           keywords={[rec.id, ...rec.labels]}
           actions={
             <ActionPanel>
-              <Action
-                title="Open in Craft Agents"
-                icon={Icon.ArrowRight}
-                onAction={() => openSession(rec.id)}
-              />
+              <Action title="Open in Craft Agents" icon={Icon.ArrowRight} onAction={() => openSession(rec.id)} />
               <Action
                 title={rec.flagged ? "Unflag" : "Flag"}
                 icon={rec.flagged ? Icon.StarDisabled : Icon.Star}
